requirements.txt
psutil~=6.0.0
azure-identity~=1.17.0
azure-cosmos~=4.7.0